Rename the power card file
