+++ src/energy-sankey.ts
@@ -1,7 +1,7 @@
 import { version } from "../package.json";
 import "./cards/energy-card/hui-energy-elec-flow-card";
 import "./cards/energy-card/energy-flow-card-editor";
-import "./cards/power-card/hui-power-flow-card";
+import "./cards/power-card/power-flow-card";
 import "./cards/power-card/power-flow-card-editor";
 import "./ha-elec-sankey";
